First WIP version of the components library
Right now is mostly ported from the ui-template repository, but using
base chakra packages instead of the global react one (doing so should
reduce a bit the output).

Depends on ChakraProvider to apply a theme, and that provider adds some
base styles that many people may not like, so this is probably gonna
change to a more generic abstraction.