Fix react effects racing condition on account change
Everything was working fine, but when you did switch accounts, an
attempt to fetch the account info when the client.wallet wasn't yet
initialized was throwing an error.

This is because when we switch accounts most of the effects detect that
account change and try to do stuff. The fetch account was checking for
client, but the error seems to be due to the wallet not being set fast
enough for the effects to take it. Everything worked fine because those
effects were executed again when the wallet was properly set to the
client.

The fix has been just to add an additional check for the client.wallet.