[major] Set env as EnvOption in state
Marked as major because the env type changed from string to its actual
type; should not affect implementations tho, since those implementations
are already casting to EnvOptions, so the result would be trying to cast
EnvOptions to EnvOptions, which will work.

Also fixed how the env prop was defined, it was using the enum keys
lowercased, rather than the actual values of the enum (should have no
effect, since the result is the same).

closes #59