Allow specify a custom id to org's fetch method
This fixes a weird case where the data of the currently connected
account could be loaded instead of the organization one.

Due to these account state changes, it's better to always specify the
id to the fetch method, but when not specified it will take it from the
state (note that, if the id is not set in the state, you will still
reproduce the aforementioned issue.